+++ lib/brasilex/ie/states/al.ex
@@ -14,6 +14,8 @@ defmodule Brasilex.IE.States.AL do
   #
   # Example: 240000048
 
+  alias Brasilex.IE.Checksum
+
   @weights [9, 8, 7, 6, 5, 4, 3, 2]
   @valid_types ~w(0 3 5 7 8)
 
@@ -21,12 +23,16 @@ defmodule Brasilex.IE.States.AL do
   Validates an Alagoas IE number (9 digits, prefix "24").
   """
   @spec validate(String.t()) :: :ok | {:error, atom()}
-  def validate(<<"24", type::binary-size(1), _rest::binary>> = digits)
-      when byte_size(digits) == 9 do
-    if type in @valid_types do
-      if valid_checksum?(digits), do: :ok, else: {:error, :invalid_checksum}
-    else
-      {:error, :invalid_format}
+  def validate(<<"24", type::binary-size(1), _::binary-size(6)>> = digits) do
+    cond do
+      type not in @valid_types ->
+        {:error, :invalid_format}
+
+      not valid_checksum?(digits) ->
+        {:error, :invalid_checksum}
+
+      true ->
+        :ok
     end
   end
 
@@ -34,26 +40,12 @@ defmodule Brasilex.IE.States.AL do
   def validate(_), do: {:error, :invalid_length}
 
   defp valid_checksum?(<<payload::binary-size(8), dv::binary-size(1)>>) do
-    calculated = calculate_dv(payload)
-    String.to_integer(dv) == calculated
-  end
-
-  defp calculate_dv(payload) do
-    sum =
-      payload
-      |> String.graphemes()
-      |> Enum.map(&String.to_integer/1)
-      |> Enum.zip(@weights)
-      |> Enum.map(fn {digit, weight} -> digit * weight end)
-      |> Enum.sum()
-
-    result = rem(sum * 10, 11)
-    if result == 10, do: 0, else: result
+    String.to_integer(dv) == Checksum.mod11_dv(payload, @weights, :rem_times_10_zero_when_10)
   end
 
   @doc """
   Formats an IE number in AL format: NNNNNNNNN
   """
   @spec format(String.t()) :: String.t()
-  def format(digits), do: digits
+  def format(digits) when byte_size(digits) == 9, do: digits
 end