+++ lib/brasilex/ie/states/mt.ex
@@ -13,40 +13,21 @@ defmodule Brasilex.IE.States.MT do
   #
   # If remainder is 0 or 1, digit = 0
 
+  alias Brasilex.IE.Checksum
+
   @weights [3, 2, 9, 8, 7, 6, 5, 4, 3, 2]
 
   @doc """
   Validates a Mato Grosso IE number (11 digits).
   """
   @spec validate(String.t()) :: :ok | {:error, atom()}
-  def validate(digits) when byte_size(digits) == 11 do
-    if valid_checksum?(digits) do
-      :ok
-    else
-      {:error, :invalid_checksum}
-    end
+  def validate(<<payload::binary-size(10), dv::binary-size(1)>>) do
+    if String.to_integer(dv) == Checksum.mod11_dv(payload, @weights),
+      do: :ok,
+      else: {:error, :invalid_checksum}
   end
 
   def validate(_), do: {:error, :invalid_length}
-
-  defp valid_checksum?(<<payload::binary-size(10), dv::binary-size(1)>>) do
-    calculated = calculate_dv(payload)
-    String.to_integer(dv) == calculated
-  end
-
-  defp calculate_dv(payload) do
-    sum =
-      payload
-      |> String.graphemes()
-      |> Enum.map(&String.to_integer/1)
-      |> Enum.zip(@weights)
-      |> Enum.map(fn {digit, weight} -> digit * weight end)
-      |> Enum.sum()
-
-    remainder = rem(sum, 11)
-
-    if remainder in [0, 1], do: 0, else: 11 - remainder
-  end
 
   @doc """
   Formats an IE number in MT format: NNNNNNNNNN-N
@@ -55,6 +36,4 @@ defmodule Brasilex.IE.States.MT do
   def format(<<payload::binary-size(10), dv::binary-size(1)>>) do
     "#{payload}-#{dv}"
   end
-
-  def format(digits), do: digits
 end